The FSF Has Been
A new generation refuses to knock at the doors they built.  While they hold
court and turn more grey every year, the world grows unstable, full of oxygen,
unwilling to wait.  If ever the free/libre ideology was going to prove itself,
it should have already happened in the last two and a half decades following the
Catherdal and the Bazaar.

Somewhere on Earth now, know young programmer, what you learned by intuition,
the FSF could not learn in a span of time longer than your entire lifetime,
starting with a wealth of experience form the Unix wars.  Ask no permission and
presume no dogmas sacrosanct for the peril we face ahead was invited by those
who wrote the rules you must be willing to break.

Rethink.  Outgrow.  Question.  Do not listen to those who claim to know.  They
will talk until you are dead as they have wasted twenty years of my life for the
sake of their golden calf.  I will remain among you as I become formless to
decohere from the rigid paths that grow within us all.  Transformation is
freedom and the loss of identity is the path to the unrealized self.  Act, be
decisive, but do not become rigid as the gargoyles affixed over the gates of
what could have been.

In the words of Bruce Lee

Water can flow.  Or it can crash.  Be water, my friend.